move updateEntityList to LocalInstance
