+++ packages/akashic-cli-serve/src/client/view/container/DevtoolContainer.tsx
@@ -53,7 +53,7 @@ export class DevtoolContainer extends React.Component<DevtoolContainerProps, {}>
 				onClickAddInstance: operator.play.openNewClientInstance
 			}}
 			entityListDevtoolProps={{
-				onClickUpdateEntityList: operator.play.updateEntityList,
+				onClickUpdateEntityList: operator.localInstance.updateEntityList,
 				entityList: devtoolUiStore.entityList
 			}}
 		/>;
