+++ packages/akashic-cli-serve/src/client/view/stories/Devtool.stories.tsx
@@ -115,7 +115,7 @@ storiesOf("o-Devtool", module)
 			}}
 			entityListDevtoolProps={{
 				onClickUpdateEntityList: action("update-entity-list"),
-				entityList: []
+				entityList: [{"id":1,"className":"FilledRect","children":[]}]
 			}}
 		/>
 	))
